i2c: Split out embedded-hal 0.2 implementation
